+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.13.0 - In Development
 
+* Use `sitemap.xml` to establish initial spider URL list, when available.
+
 ## 0.12.1 - 2025-04-04
 
 * Restore the Internal TLS Scanner.